S08: quantifiers, alternation and anchors, 693 corpus rows matching upstream
Parsing/ gains Branch and every one of its optimise helpers, GreedyRepeat,
LazyRepeat and PossessiveRepeat, the sixteen ZeroWidthBase subclasses, and the
parser side of all three: apply_quantifier, parse_quantifier, is_above_limit,
parse_limited_quantifier, parse_count, the ^/$ and ?*+{ branches of
parse_sequence, and the positional escapes \A \Z \b \B \m \M \G \K.

Ratchet green at 786 passing tests, up from 514. The compile-parity corpus goes
from 436 to 693 of 1547 compile rows and 16 to 23 of 50 error rows, with no row
failing. quantifiers falls from 460 waiting tests to 179, anchors from 261 to
87, and alternation leaves the board; character-classes is now the largest block
at 573, which is S10.

Most of this landed in a driver session that the machine killed mid-flight, with
two claims written down before their evidence had arrived. Both are now settled
against the real Python module, regex 2026.7.19:

  .??? really is "multiple repeat" at position 3. The ported test asserted the
  message "multiple repeat at position 3", which is Python's str(error) and not
  this port's shape - S01 split msg and pos apart - so it was red. It now asserts
  the message and the Offset separately, which is stricter than upstream's own
  assertRaisesRegex against the bare string "multiple repeat". The other three
  tests in that file carry the same Phase 1 transcription and each needs the same
  split when its slice turns it on.

  max_width saturation at long.MaxValue is indistinguishable from upstream's
  unbounded int, confirmed by bytecode rather than by argument. Saturation can
  only happen above long.MaxValue, which is 2^31 times UNLIMITED, and every
  consumer either takes min(w, UNLIMITED) or asks w >= UNLIMITED.

Getting that second answer needed a technique worth keeping. Upstream's C
compiler is O(repeat count), so regex.compile('a{4294967294}') runs for about
twenty minutes and then raises MemoryError, while everything before it is O(1) -
parse, optimise, compile and flatten all measure 0.000s at a count of 8,000,000.
Capturing the arguments to regex._regex.compile and not calling through returns
in milliseconds. The dead session's script called through, which is the whole
reason it looked like a hang.

Gaps/Parsing/RepeatWidthOverflowTests.cs pins the result: the saturating case,
a non-saturating control that would catch a Multiply that always saturated, and
a single maximal count. Writing it found a real bug. CompiledPattern.ReqOffset
was an int, and an offset is a repeat's maximum width, which runs to
UNLIMITED - 1 = 4294967294 - so (?:a{65535}){0,65535}a's offset of 4294836225
came back as -131071. It is a long end to end now. No corpus row has an offset
above int.MaxValue, which is why 693 rows passed with the bug in place.

The blind pass raised nothing, and did it differentially rather than by eye:
about 34,000 (pattern, flags) pairs compiled on both sides, 19,000 of them
reaching real bytecode, comparing code, req_offset, req_chars, req_flags,
group_count and the resolved flags; hand-built waves over every construct in
scope, six seeds of grammar fuzzing at depth 3, ~150 maximal-count shapes that
independently confirm the three expectations in the new test, and a negative
control proving the comparator detects an injected wrong opcode. The four hunt
items - dict order, set order, structural equality, UNLIMITED versus None - were
cleared by reading as well: every ordering leak that reaches the bytecode is
still behind a needs: seam, the node equalities match upstream's _key, and
MaxCount stays nullable everywhere. No second pass was needed; no code changed
after the review.

One out-of-scope difference it noticed is left alone: (?V0:x(?V1:y)) raises
KeyError: regex.V0|V1 upstream but ArgumentOutOfRangeException here, from
RegexFlags.cs, which S07 wrote and this diff does not touch. Both are
non-error exceptions, so no corpus row can see it.

+++ src/FuzzyRegex/Parsing/CompiledPattern.cs
@@ -39,6 +39,10 @@ namespace Fuzzy.Text.RegularExpressions.Parsing;
 /// </param>
 /// <param name="ReqOffset">
 /// Where the required string starts, or -1 when the offset is unbounded, or 0 when there is none.
+/// A <c>long</c> because the offset is a repeat's maximum width and a repeat count runs to
+/// <c>UNLIMITED - 1</c> = 4294967294, which does not fit in an <see cref="int"/>: upstream hands
+/// the C compiler a Python <c>int</c> that it reads as a <c>Py_ssize_t</c>
+/// (<c>upstream/src/_regex.c</c>). Pinned by <c>Gaps/Parsing/RepeatWidthOverflowTests.cs</c>.
 /// </param>
 /// <param name="ReqChars">The required string, as codepoints, after case folding.</param>
 /// <param name="ReqFlags">The case flags the required string must be compared under.</param>
@@ -49,7 +53,7 @@ internal sealed record CompiledPattern(
     IReadOnlyDictionary<string, int> GroupIndex,
     IReadOnlyDictionary<string, IReadOnlySet<string>> NamedLists,
     IReadOnlyList<IReadOnlySet<string>> NamedListIndexes,
-    int ReqOffset,
+    long ReqOffset,
     IReadOnlyList<int> ReqChars,
     int ReqFlags,
     int GroupCount
